% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_na_to.R
\name{convert_na_to}
\alias{convert_na_to}
\alias{convert_na_to.numeric}
\alias{convert_na_to.data.frame}
\title{Replace missing values in a variable or a dataframe.}
\usage{
convert_na_to(x, ...)

\method{convert_na_to}{numeric}(x, replacement = NULL, verbose = TRUE)

\method{convert_na_to}{data.frame}(
  x,
  replacement_num = NULL,
  replacement_char = NULL,
  replacement_fac = NULL,
  select = NULL,
  exclude = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric, factor, or character vector, or a data frame.}

\item{...}{Not used.}

\item{replacement}{Numeric or character value that will be used to replace \code{NA}.}

\item{verbose}{Toggle warnings.}

\item{select}{Character vector of column names. If \code{NULL} (the default), all
variables will be selected.}

\item{exclude}{Character vector of column names to be excluded from selection.}
}
\value{
\code{x}, where all \code{NA} values are replaced by \code{replacement}.
}
\description{
Replace missing values in a variable or a dataframe.
}
