% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_arrange.R
\name{data_arrange}
\alias{data_arrange}
\alias{data_arrange.default}
\title{Arrange rows by column values}
\usage{
data_arrange(data, ...)

\method{data_arrange}{default}(data, select = NULL, by = NULL, safe = TRUE, ...)
}
\arguments{
\item{data}{A (grouped) data frame, or an object that can be coerced to a
data frame.}

\item{...}{Currently not used.}

\item{select}{Character vector of column names. Use a dash just before column
name to arrange in decreasing order, for example \code{"-x1"}.}

\item{by}{Optional character string, indicating the names of one or more
variables in the data frame. If supplied, data rows will be sorted within
each group.}

\item{safe}{Do not throw an error if one of the variables specified doesn't
exist.}
}
\value{
A data frame, where rows are sorted according to the selected columns.
}
\description{
\code{data_arrange()} orders the rows of a data frame by the values of selected
columns.
}
\examples{

# Arrange using several variables
data_arrange(head(mtcars), c("gear", "carb"))

# Arrange in decreasing order
data_arrange(head(mtcars), "-carb")

# compare: arrange variables vs. arrange variable within groups
set.seed(123)
sample_rows <- sample(seq_len(nrow(iris)), 10, replace = TRUE)
x <- iris[sample_rows, c("Sepal.Width", "Species")]
data_arrange(x, c("Sepal.Width", "Species"))
data_arrange(x, "Sepal.Width", by = "Species")

# Throw an error if one of the variables specified doesn't exist
try(data_arrange(head(mtcars), c("gear", "foo"), safe = FALSE))
}
