% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_arrange2.R
\name{data_arrange2}
\alias{data_arrange2}
\title{Arrange rows by column values}
\usage{
data_arrange2(
  data,
  select = NULL,
  exclude = NULL,
  ascending = NULL,
  descending = NULL,
  ignore_case = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame, or an object that can be coerced to a data frame.}

\item{...}{Quoted column names. Use a dash just before column name to arrange
in decreasing order, for example "-x1".}

\item{safe}{Do not throw an error if one of the variables specified doesn't
exist.}
}
\value{
A data frame.
}
\description{
\code{data_arrange2()} orders the rows of a data frame by the values of selected
columns.
}
\examples{
# Arrange using several variables
data_arrange(head(mtcars), "gear", "carb")

# Arrange in decreasing order
data_arrange(head(mtcars), "-carb")

# Throw an error if one of the variables specified doesn't exist
data_arrange(head(mtcars), "gear", "foo", safe = FALSE)

}
