% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_duplicated.R
\name{data_duplicated}
\alias{data_duplicated}
\title{Extract all duplicates}
\usage{
data_duplicated(data, select)
}
\arguments{
\item{data}{The data frame.}

\item{select}{The id variable for which to check for duplicates.}
}
\value{
A dataframe, containing all duplicates.
}
\description{
Extract all duplicates, for visual inspection.
Note that it also contains the first occurrence of future
duplicates, unlike \code{\link[=duplicated]{duplicated()}} or \code{\link[dplyr:distinct]{dplyr::distinct()}}). Also
contains an additional column reporting the number of missing
values for that row, to help in the decision-making when
selecting which duplicates to keep.
}
\examples{
df1 <- data.frame(
   id = c(1, 2, 3, 1, 3),
   year = c(2022, 2022, 2022, 2022, 2000),
   item1 = c(NA, 1, 1, 2, 3),
   item2 = c(NA, 1, 1, 2, 3),
   item3 = c(NA, 1, 1, 2, 3)
)

data_duplicated(df1, select = "id")

data_duplicated(df1, select = c("id", "year"))

# Filter to exclude duplicates
df2 <- df1[-c(1, 5),]
df2
}
\seealso{
\code{\link[=data_unique]{data_unique()}}
}
\keyword{duplicates}
