% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge.R
\name{data_merge}
\alias{data_merge}
\alias{data_merge.data.frame}
\title{Merge (join) two data frames, or a list of data frames}
\usage{
data_merge(x, ...)

\method{data_merge}{data.frame}(x, y, join = "left", by = NULL, id = NULL, verbose = TRUE, ...)
}
\arguments{
\item{x, y}{A data frame to merge}

\item{...}{Not used.}

\item{join}{Character vector, indicating the way of joining the data frames.}

\item{by}{Specifications of the columns used for merging.}

\item{id}{Optional name for ID column that will be created to indicate the source data frames for appended rows.}

\item{verbose}{Toggle warnings.}
}
\value{
A merged data frame.
}
\description{
Merge (join) two data frames, or a list of data frames.
}
\examples{
data(iris)
}
