% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{data_read}
\alias{data_read}
\title{Read (import) data files from various sources}
\usage{
data_read(path, path_catalog = NULL, encoding = NULL, verbose = TRUE, ...)
}
\arguments{
\item{path}{Character string, the file path to the data file.}

\item{path_catalog}{Character string, path to the catalog file. Only relevant
for SAS data files.}

\item{encoding}{The character encoding used for the file. Usually not needed.}

\item{verbose}{Toggle warnings and messages.}

\item{...}{Arguments passed to the related \verb{read_*()} function.}
}
\value{
A data frame.
}
\description{
This functions imports data from various file types. This function is a small
wrapper around \code{haven::read_spss()}, \code{haven::read_stata()}, \code{haven::read_sas()},
\code{readxl::read_excel()} and \code{data.table::fread()} resp. \code{readr::read_delim()}
(the latter if package \strong{data.table} is not installed). Thus, supported file
types for importing data are data files from SPSS, SAS or Stata, Excel files
or text files (like '.csv' files).
}
\details{
\code{data_read()} detects the appropriate \verb{read_*()} function based on
the file-extension of the data file. For data files from SPSS, SAS or Stata,
variables are converted into their most appropriate type, i.e. variables
where \emph{all} values are labelled, will be converted into factors, where imported
value labels will be set as factor levels. Else, if a variable has \emph{no} value
labels or less value labels than values, the variable is either converted
into numeric or character vector. Value labels are then preserved as \code{"labels"}
attribute.
}
