% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recode.R
\name{data_recode}
\alias{data_recode}
\alias{data_recode.numeric}
\title{Recode data}
\usage{
data_recode(x, ...)

\method{data_recode}{numeric}(x, split = "median", n_groups = NULL)
}
\arguments{
\item{x}{A data frame or vector.}

\item{split}{Name of a function, or numeric values, indicating cutoffs}

\item{n_groups}{If \code{split = "quantile"}.}
}
\description{
Recode data.
}
