% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_findcols.R, R/data_remove.R,
%   R/data_reorder.R
\name{data_findcols}
\alias{data_findcols}
\alias{data_remove}
\alias{data_reorder}
\title{Find, reorder or remove specific columns}
\usage{
data_findcols(data, pattern = NULL, starts_with = NULL, ends_with = NULL, ...)

data_remove(data, pattern, ...)

data_reorder(data, cols, safe = TRUE, ...)
}
\arguments{
\item{data}{A data frame, or an object that can be coerced to a data frame.}

\item{pattern, starts_with, ends_with}{Character strings.}

\item{...}{Other arguments passed to or from other functions.}

\item{cols}{Vector of column names.}

\item{safe}{Do not throw error if for instance the variable to be
renamed/removed doesn't exist.}
}
\description{
Safe and intuitive functions to reorder or remove variables in dataframes.
}
\examples{
# Find columns names by pattern
data_findcols(iris, starts_with = "Sepal")
data_findcols(iris, ends_with = "Width")
data_findcols(iris, pattern = "\\\\.")
data_findcols(iris, c("Petal.Width", "Sepal.Length"))
# Remove columns
head(data_remove(iris, "Sepal.Length"))
# Reorder columns
head(data_reorder(iris, c("Species", "Sepal.Length")))
head(data_reorder(iris, c("Species", "dupa"))) # Safe for non-existing cols

}
\seealso{
\itemize{
\item Functions to rename stuff: \code{\link[=data_rename]{data_rename()}}, \code{\link[=data_rename_rows]{data_rename_rows()}}, \code{\link[=data_addprefix]{data_addprefix()}}, \code{\link[=data_addsuffix]{data_addsuffix()}}
\item Functions to reorder, find and remove columns: \code{\link[=data_findcols]{data_findcols()}}, \code{\link[=data_reorder]{data_reorder()}}, \code{\link[=data_remove]{data_remove()}}
\item Functions to reshape, pivot or rotate dataframes: \code{\link[=data_to_long]{data_to_long()}}, \code{\link[=data_to_wide]{data_to_wide()}}, \code{\link[=data_rotate]{data_rotate()}}
\item Functions to rescale and reverse: \code{\link[=data_rescale]{data_rescale()}}, \code{\link[=data_reverse]{data_reverse()}}
\item Functions to standardize, normalize, rank-transform: \code{\link[=standardize]{standardize()}}, \code{\link[=normalize]{normalize()}}, \code{\link[=ranktransform]{ranktransform()}}
\item Split, cut and merge dataframes: \code{\link[=data_partition]{data_partition()}}, \code{\link[=data_cut]{data_cut()}}, \code{\link[=data_match]{data_match()}}, \code{\link[=data_merge]{data_merge()}}
}
}
