% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rotate.R
\name{data_rotate}
\alias{data_rotate}
\title{Rotate a dataframe}
\usage{
data_rotate(data, rownames = NULL, colnames = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{A dataframe.}

\item{rownames}{Character vector (optional). If not \code{NULL}, the dataframe's
rownames will be added as (first) column to the output, with \code{rownames}
being the name of this column.}

\item{colnames}{Logical (optional), if \code{TRUE}, the values of the first column
in \code{x} will be used as column names in the rotated dataframe.}

\item{verbose}{Toggle warnings.}
}
\value{
A (rotated) dataframe.
}
\description{
This function rotates a dataframe, i.e. columns become rows and vice versa.
}
\examples{
x <- mtcars[1:3, 1:4]
data_rotate(x)
data_rotate(x, rownames = "property")

# use values in 1. column as column name
data_rotate(x, colnames = TRUE)
data_rotate(x, rownames = "property", colnames = TRUE)
}
