% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{data_transform}
\alias{data_transform}
\title{Create new variables in a data frame}
\usage{
data_transform(data, ...)
}
\arguments{
\item{data}{The data frame, for which new variables should be created.}

\item{...}{A sequence of named expressions, where the left-hand side refers
to the name of the new variable, while the right-hand side represent the
values of the new variable. See 'Examples'.}
}
\description{
Create new variables in a data frame. Unlike \code{base::transform()}, \code{data_transform}
can be used on grouped data frames, and newly created variables can be directly
used.
}
\note{
This function is still experimental and well tested for interactive
use. There might be corner cases, e.g. when called from inside functions
or similar, where \code{data_transform()} does not yet work. Please carefully
check your results before using this function, say, in package code.
}
\examples{
data(efc)
new_efc <- data_transform(
  efc,
  c12hour_c = center(c12hour),
  c12hour_z = c12hour_c / sd(c12hour, na.rm = TRUE),
  c12hour_z2 = standardize(c12hour)
)
head(new_efc)

# attributes - in this case, value and variable labels - are preserved
str(new_efc)

# works on grouped data
grouped_efc <- data_group(efc, "c172code")
new_efc <- data_transform(
  grouped_efc,
  c12hour_c = center(c12hour),
  c12hour_z = c12hour_c / sd(c12hour, na.rm = TRUE),
  c12hour_z2 = standardize(c12hour)
)
head(new_efc)
}
