% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_missing.R
\name{describe_missing}
\alias{describe_missing}
\title{Describe Missing Values in Data According to Guidelines}
\usage{
describe_missing(data, vars = NULL, scales = NULL)
}
\arguments{
\item{data}{The data frame to be analyzed.}

\item{vars}{Variable (or lists of variables) to check for missing values (NAs).}

\item{scales}{The scale names to check for missing values (as a character vector).}
}
\value{
A dataframe with the following columns:
\itemize{
\item \code{var}: Variables selected.
\item \code{items}: Number of items for selected variables.
\item \code{na}: Number of missing cell values for those variables (e.g., 2 missing
values for the first participant + 2 missing values for the second participant
= total of 4 missing values).
\item \code{cells}: Total number of cells (i.e., number of participants multiplied by
the number of variables, \code{items}).
\item \code{na_percent}: The percentage of missing values (\code{na} divided by \code{cells}).
\item \code{na_max}: The number of missing values for the participant with the most
missing values for the selected variables.
\item \code{na_max_percent}: The amount of missing values for the participant with
the most missing values for the selected variables, as a percentage
(i.e., \code{na_max} divided by the number of selected variables, \code{items}).
\item \code{all_na}: The number of participants missing 100\% of items for that scale
(the selected variables).
}
}
\description{
Provides a detailed description of missing values in a data frame.
This function reports both absolute and percentage missing values of specified
column lists or scales, following recommended guidelines. Some authors recommend
reporting item-level missingness per scale, as well as a participant's maximum
number of missing items by scale. For example, Parent (2013) writes:

\emph{I recommend that authors (a) state their tolerance level for missing data by scale
or subscale (e.g., "We calculated means for all subscales on which participants gave
at least 75\% complete data") and then (b) report the individual missingness rates
by scale per data point (i.e., the number of missing values out of all data points
on that scale for all participants) and the maximum by participant (e.g., "For Attachment
Anxiety, a total of 4 missing data points out of 100 were observed, with no participant
missing more than a single data point").}
}
\examples{
# Use the entire data frame
describe_missing(airquality)

# Use selected columns explicitly
describe_missing(airquality,
  vars = list(
    c("Ozone", "Solar.R", "Wind"),
    c("Temp", "Month", "Day")
  )
)

# If the questionnaire items start with the same name, e.g.,
set.seed(15)
fun <- function() {
  c(sample(c(NA, 1:10), replace = TRUE), NA, NA, NA)
}
df <- data.frame(
  ID = c("idz", NA),
  open_1 = fun(), open_2 = fun(), open_3 = fun(),
  extrovert_1 = fun(), extrovert_2 = fun(), extrovert_3 = fun(),
  agreeable_1 = fun(), agreeable_2 = fun(), agreeable_3 = fun()
)

# One can list the scale names directly:
describe_missing(df, scales = c("ID", "open", "extrovert", "agreeable"))
}
\references{
Parent, M. C. (2013). Handling item-level missing
data: Simpler is just as good. \emph{The Counseling Psychologist},
\emph{41}(4), 568-600. https://doi.org/10.1177\%2F0011000012445176
}
\keyword{NA}
\keyword{guidelines}
\keyword{missing}
\keyword{values}
