% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_deprecated}
\alias{.is_deprecated}
\title{Print a message saying that an argument is deprecated and that the user
should use its replacement instead.}
\usage{
.is_deprecated(arg, replacement)
}
\arguments{
\item{arg}{Argument that is deprecated}

\item{replacement}{Argument that replaces the deprecated argument}
}
\description{
Print a message saying that an argument is deprecated and that the user
should use its replacement instead.
}
\keyword{internal}
