% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rescale.R
\name{.update_to}
\alias{.update_to}
\title{expand the new target range by multiplying or adding}
\usage{
.update_to(x, to, multiply, add)
}
\description{
expand the new target range by multiplying or adding
}
\keyword{internal}
