% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\alias{median_mad}
\title{Summary Helpers}
\usage{
mean_sd(x, na.rm = TRUE, named = TRUE, ...)

median_mad(x, na.rm = TRUE, constant = 1.4826, named = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector (or one that can be coerced to one via
\code{as.numveric()}) to be summarized.}

\item{na.rm}{if \code{TRUE} then \code{NA} values are stripped
    from \code{x} before computation takes place.}

\item{...}{Not used.}

\item{constant}{scale factor.}
}
\value{
A vector of length 3: one SD below the mean, the mean, and one SD
above the mean (or median and MAD).
}
\description{
Summary Helpers
}
\examples{
mean_sd(mtcars$mpg)

median_mad(mtcars$mpg)

}
