% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rows.R
\name{rownames_as_column}
\alias{rownames_as_column}
\alias{column_as_rownames}
\alias{rowid_as_column}
\title{Tools for working with row names or row ids}
\usage{
rownames_as_column(x, var = "rowname")

column_as_rownames(x, var = "rowname")

rowid_as_column(x, var = "rowid")
}
\arguments{
\item{x}{A data frame.}

\item{var}{Name of column to use for row names/ids. For \code{column_as_rownames()},
this argument can be the variable name or the column number.}
}
\value{
A data frame.
}
\description{
Tools for working with row names or row ids
}
\examples{
# Convert between row names and column --------------------------------
test <- rownames_as_column(mtcars, var = "car")
test
head(column_as_rownames(test, var = "car"))

test_data <- head(iris)

rowid_as_column(test_data)
rowid_as_column(test_data, var = "my_id")
}
