% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seek_variables.R
\name{seek_variables}
\alias{seek_variables}
\title{Find variables by its name, variable or value labels}
\usage{
seek_variables(data, pattern, search = c("names", "labels"), fuzzy = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{pattern}{Character string (regular expression) to be matched in \code{data}.
May also be a character vector of length > 1. \code{pattern} is searched for in
column names, variable label and value labels attributes, or factor levels of
variables in \code{data}.}

\item{search}{Character vector, indicating where \code{pattern} is sought. Use one
of more of the following options:
\itemize{
\item \code{"name"}: searches in column names.
\item \code{"labels"}: searches in variable labels. Only applies when a \code{label} attribute
is set for a variable.
\item \code{"values"}: searches in value labels or factor levels. Only applies when a
\code{labels} attribute is set for a variable, or if a variable is a factor.
\item \code{"all"}: searches in all of the above.
}}

\item{fuzzy}{Logical, if \code{TRUE}, "fuzzy matching" (partial and close distance
matching) will be used to find \code{pattern}.}
}
\value{
A data frame with three columns: the column index, the column name
and - if available - the variable label of all matched variables in \code{data}.
}
\description{
This functions seeks variables in a data frame, based on patterns
that either match the variable name (column name), variable labels, value labels
or factor levels. Matching variable and value labels only works for "labelled"
data, i.e. when the variables either have a \code{label} attribute or \code{labels}
attribute.

\code{seek_variables()} is particular useful for larger data frames with labelled
data - finding the correct variable name can be a challenge. This function
helps to find the required variables, when only certain patterns of variable
names or labels are known.
}
\examples{
# seek variables with "Length" in variable name or labels
seek_variables(iris, "Length")

# seek variables with "dependency" in names or labels
# column "e42dep" has a label-attribute "elder's dependency"
data(efc)
seek_variables(efc, "dependency")

# "female" only appears as value label attribute - default search is in
# variable names and labels only, so no match
seek_variables(efc, "female")
# when we seek in all sources, we find the variable "e16sex"
seek_variables(efc, "female", search = "all")

# typo, no match
seek_variables(iris, "Lenght")
# typo, fuzzy match
seek_variables(iris, "Lenght", fuzzy = TRUE)
}
