% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.R
\name{smoothness}
\alias{smoothness}
\title{Series smoothness}
\usage{
smoothness(x, method = "cor", lag = 1, iterations = NULL, ...)
}
\arguments{
\item{x}{Numeric vector (similar to a time series).}

\item{method}{Can be \code{"diff"} (the standard deviation of the standardized
differences) or \code{"cor"} (default, lag-one autocorrelation).}

\item{lag}{An integer indicating which lag to use. If less than \code{1}, will be
interpreted as expressed in percentage of the length of the vector.}

\item{iterations}{The number of bootstrap replicates for computing standard
errors. If \code{NULL} (default), parametric standard errors are computed.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
Value of smoothness.
}
\description{
Functions to quantify the smoothness of a vector, which can be used in some cases
as an index of "linearity". A smooth series is one that does not have abrupt changes in
its values. The smoothness of a series can be approximated in different ways, such
as the standard deviation of the standardized differences or the lag-one
autocorrelation.
}
\examples{
x <- (-10:10)^3 + rnorm(21, 0, 100)
plot(x)
smoothness(x, method = "cor")
smoothness(x, method = "diff")

# A bootstrapped value can also be computed
smoothness(x, iterations = 100)

# When perfectly linear, the "smoothness" is 1
smoothness(1:10)

# And closer to zero for random
smoothness(rnorm(1000))
smoothness(rnorm(1000), method = "diff")

}
\references{
https://stats.stackexchange.com/questions/24607/how-to-measure-smoothness-of-a-time-series-in-r
}
