% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R, R/unstandardize.R
\name{standardize}
\alias{standardize}
\alias{unstandardize}
\title{Standardization (Z-scoring)}
\usage{
standardize(
  x,
  robust = FALSE,
  two_sd = FALSE,
  weights = NULL,
  verbose = TRUE,
  ...
)

unstandardize(
  x,
  center = NULL,
  scale = NULL,
  reference = NULL,
  robust = FALSE,
  two_sd = FALSE,
  ...
)
}
\arguments{
\item{x}{A data frame, a vector or a statistical model (for `unstandardize()`
cannot be a model).}

\item{robust}{Logical, if `TRUE`, centering is done by subtracting the
median from the variables and dividing it by the median absolute deviation
(MAD). If `FALSE`, variables are standardized by subtracting the
mean and dividing it by the standard deviation (SD).}

\item{two_sd}{If `TRUE`, the variables are scaled by two times the deviation
(SD or MAD depending on `robust`). This method can be useful to obtain
model coefficients of continuous parameters comparable to coefficients
related to binary predictors, when applied to **the predictors** (not the
outcome) (Gelman, 2008).}

\item{weights}{Can be `NULL` (for no weighting), or:
- For model: if `TRUE` (default), a weighted-standardization is carried out.
- For `data.frame`s: a numeric vector of weights, or a character of the
  name of a column in the `data.frame` that contains the weights.
- For numeric vectors: a numeric vector of weights.}

\item{verbose}{Toggle warnings and messages on or off.}

\item{...}{Arguments passed to or from other methods.}

\item{center, scale}{Used by `unstandardize()`; `center` and `scale`
correspond to the center (the mean / median) and the scale (SD / MAD) of
the original non-standardized data (for data frames, should be named, or
have column order correspond to the numeric column). However, one can also
directly provide the original data through `reference`, from which the
center and the scale will be computed (according to `robust` and `two_sd`.
Alternatively, if the input contains the attributes `center` and `scale`
(as does the output of `standardize()`), it will take it from there if the
rest of the arguments are absent.}
}
\value{
The standardized object (either a standardize data frame or a
  statistical model fitted on standardized data).
}
\description{
Performs a standardization of data (z-scoring), i.e., centering and scaling,
so that the data is expressed in terms of standard deviation (i.e., mean = 0,
SD = 1) or Median Absolute Deviance (median = 0, MAD = 1). When applied to a
statistical model, this function extracts the dataset, standardizes it, and
refits the model with this standardized version of the dataset. The
[normalize()] function can also be used to scale all numeric variables within
the 0 - 1 range.
}
\note{
When `x` is a vector or a data frame with `remove_na = "none")`,
  missing values are preserved, so the return value has the same length /
  number of rows as the original input.
}
\examples{
# Data frames
summary(standardize(swiss))

# Models
model <- lm(Infant.Mortality ~ Education * Fertility, data = swiss)
coef(standardize(model))
}
\seealso{
Other transform utilities: 
\code{\link{normalize}()}
}
\concept{standardize}
\concept{transform utilities}
