% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_at.R
\name{value_at}
\alias{value_at}
\title{Find the value(s) at a specific position in a variable}
\usage{
value_at(x, position = 1, default = NULL, remove_na = FALSE)
}
\arguments{
\item{x}{A vector or factor.}

\item{position}{An integer or a vector of integers, indicating the position(s)
of the value(s) to be returned. Negative values are counted from the end of
the vector. If \code{NA}, an error is thrown.}

\item{default}{The value to be returned if the position is out of range.}

\item{remove_na}{Logical, if \code{TRUE}, missing values are removed before
computing the position. If \code{FALSE}, missing values are included in the
computation.}
}
\value{
A vector with the value(s) at the specified position(s).
}
\description{
This function can be used to extract one or more values at a
specific position in a variable.
}
\examples{
data(mtcars)
# 5th value
value_at(mtcars$mpg, 5)
# last value
value_at(mtcars$mpg, -1)
# out of range, return default
value_at(mtcars$mpg, 150)
# return 2nd and fifth value
value_at(mtcars$mpg, c(2, 5))
}
\seealso{
\code{data_summary()} to use \code{value_at()} inside a \code{data_summary()} call.
}
