% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tabulate.R
\name{as.prop.table}
\alias{as.prop.table}
\alias{as.prop.table.datawizard_crosstab}
\title{Convert a crosstable to a propensity table}
\usage{
as.prop.table(x, ...)

\method{as.prop.table}{datawizard_crosstab}(x, remove_na = TRUE, simplify = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{datawizard_crosstab}.}

\item{...}{not used.}

\item{remove_na}{Logical, if \code{FALSE}, missing values are included in the
frequency or crosstable, else missing values are omitted. Note that the
default for the \code{as.table()} method is \code{remove_na = TRUE}, so that missing
values are not included in the returned table, which makes more sense for
post-processing of the table, e.g. using \code{chisq.test()}.}

\item{simplify}{Logical, if \code{TRUE}, the returned table is simplified to a
single table object if there is only one frequency or contingency table
input. Else, always for multiple table inputs or when \code{simplify = FALSE}, a
list of tables is returned. This is only relevant for the \code{as.table()}
methods. To ensure consistent output, the default is \code{FALSE}.}

\item{verbose}{Toggle warnings and messages.}
}
\description{
This is an S3 generic. It can be used on objects of class
\code{datawizard_crosstab} created by \code{data_tabulate()} when it was run with the
argument \code{by}.
}
\examples{
data(efc)

# Some cross tabulation
cross <- data_tabulate(efc, select = "e42dep", by = "c172code", proportions = "row")
cross

# We convert it to a propensity table
as.prop.table(cross)
}
\seealso{
\link{data_tabulate}
}
