% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modify.R
\name{data_modify}
\alias{data_modify}
\title{Create new variables in a data frame}
\usage{
data_modify(data, ...)
}
\arguments{
\item{data}{The data frame, for which new variables should be created.}

\item{...}{A sequence of named expressions, where the left-hand side refers
to the name of the new variable, while the right-hand side represent the
values of the new variable. The right-hand side can also be represented
as character string. See 'Examples'.}

\item{verbose}{Toggle messages.}
}
\description{
Create new variables in a data frame. Unlike \code{base::transform()}, \code{data_modify()}
can be used on grouped data frames, and newly created variables can be directly
used.
}
\note{
This function is still experimental and well tested for interactive
use. There might be corner cases, e.g. when called from inside functions
or similar, where \code{data_modify()} does not yet work. Please carefully
check your results before using this function, say, in package code.
}
\examples{
data(efc)
new_efc <- data_modify(
  efc,
  c12hour_c = center(c12hour),
  c12hour_z = c12hour_c / sd(c12hour, na.rm = TRUE),
  c12hour_z2 = standardize(c12hour)
)
head(new_efc)

# using character strings
new_efc <- data_modify(
  efc,
  c12hour_c = center(c12hour),
  c12hour_z = "c12hour_c / sd(c12hour, na.rm = TRUE)"
)
head(new_efc)

# attributes - in this case, value and variable labels - are preserved
str(new_efc)

# works on grouped data
grouped_efc <- data_group(efc, "c172code")
new_efc <- data_modify(
  grouped_efc,
  c12hour_c = center(c12hour),
  c12hour_z = c12hour_c / sd(c12hour, na.rm = TRUE),
  c12hour_z2 = standardize(c12hour)
)
head(new_efc)
}
