% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recode.R
\name{data_recode}
\alias{data_recode}
\alias{data_recode.numeric}
\alias{data_recode.data.frame}
\title{Recode data}
\usage{
data_recode(x, ...)

\method{data_recode}{numeric}(
  x,
  split = "median",
  n_groups = NULL,
  size_groups = NULL,
  lowest = 1,
  ...
)

\method{data_recode}{data.frame}(
  x,
  split = "median",
  n_groups = NULL,
  size_groups = NULL,
  lowest = 1,
  select = NULL,
  exclude = NULL,
  force = FALSE,
  ...
)
}
\arguments{
\item{x}{A data frame or vector.}

\item{...}{not used.}

\item{split}{Name of a function, or numeric values, indicating cutoffs}

\item{n_groups}{If \code{split = "quantile"}.}

\item{size_groups}{If \code{split = "range"}.}

\item{lowest}{Minimum value if numeric variables are recoded.}

\item{select}{Character vector of column names. If \code{NULL} (the default), all
variables will be selected.}

\item{exclude}{Character vector of column names to be excluded from selection.}

\item{force}{Logical, if \code{TRUE}, forces recoding of factors and dates
as well.}
}
\description{
Recode data.
}
\examples{
set.seed(123)
x <- sample(1:10, size = 50, replace = TRUE)

table(x)

# by default, at median
table(data_recode(x))

# into 3 groups, based on distribution (quantiles)
table(data_recode(x, split = "quantile", n_groups = 3))

# into 3 groups, try to return similar group sizes
# (i.e. similar count for each value/level)
table(data_recode(x, split = "quantile", n_groups = 3))

# into 3 groups, manual cut offs
table(data_recode(x, split = c(3, 5)))

set.seed(123)
x <- sample(1:100, size = 500, replace = TRUE)

# into 5 groups, try to return similar group sizes
table(data_recode(x, split = "equal_size", n_groups = 5))

# into 5 groups, try to return same range within groups
# i.e. 1-20, 21-40, 41-60, etc.
table(data_recode(x, split = "equal_range", size_groups = 20))
}
