% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recode.R
\name{data_recode}
\alias{data_recode}
\alias{data_recode.numeric}
\alias{data_recode.data.frame}
\title{Recode data}
\usage{
data_recode(x, ...)

\method{data_recode}{numeric}(
  x,
  split = "median",
  n_groups = NULL,
  range = NULL,
  lowest = 1,
  ...
)

\method{data_recode}{data.frame}(
  x,
  split = "median",
  n_groups = NULL,
  range = NULL,
  lowest = 1,
  select = NULL,
  exclude = NULL,
  force = FALSE,
  append = FALSE,
  ...
)
}
\arguments{
\item{x}{A data frame, numeric vector or factor.}

\item{...}{not used.}

\item{split}{Character vector, indicating where to split variables, or
numeric values, indicating cut-off values. If character, may be one of
\code{"median"}, \code{"mean"}, \code{"quantile"}, \code{"equal_size"} or \code{"equal_range"}.}

\item{n_groups}{If \code{split} is \code{"quantile"} (or its alias \code{"equal_size"}),
this defines the number of requested groups (i.e. resulting number of
levels or values) for the recoded variable(s).}

\item{range}{If \code{split = "equal_range"}, this defines the range of values
that are recoded into a new value.}

\item{lowest}{Minimum value of the recoded variable(s).}

\item{select}{Character vector of column names. If \code{NULL} (the default), all
variables will be selected.}

\item{exclude}{Character vector of column names to be excluded from selection.}

\item{force}{Logical, if \code{TRUE}, forces recoding of factors as well.}

\item{append}{Logical or string. If \code{TRUE}, recoded variables get new
column names (with the suffix \code{"_r"}) and are appended (column bind) to \code{x},
thus returning both the original and the recoded variables. If \code{FALSE},
original variables in \code{x} will be overwritten by their recoded versions.
If a character value, recoded variables are appended with new column
names (using the defined suffix) to the original data frame.}
}
\description{
Recode data.
}
\details{
\subsection{Splits and cut-off values}{
Cut-off values, which are based on the median, mean or quantile
functions, are \emph{exclusive}, this means that these values indicate the
lower bound of the next group to begin. Take a simple example, a numeric
variable with values from 1 to 9. The median would be 5, thus 1-4 are
recoded into 1, while 5-9 would turn into 2 (compare
\code{cbind(1:9, data_recode(1:9))}). The same variable, using \code{split = "quantile"}
and \code{n_groups = 3} would define cut-off points at 3.67 and 6.33 (see
\verb{quantile(1:9, probs = c(1/3, 2/3)}), which means that values from 1 to 3
are recoded into 1 (because the next group of values to be recoded starts
at 3.67), 4 to 6 into 2 and 7 to 9 into 3.
}

\subsection{Recoding into groups with equal range}{
}
}
\examples{
set.seed(123)
x <- sample(1:10, size = 50, replace = TRUE)

table(x)

# by default, at median
table(data_recode(x))

# into 3 groups, based on distribution (quantiles)
table(data_recode(x, split = "quantile", n_groups = 3))

# into 3 groups, manual cut offs
table(data_recode(x, split = c(3, 5)))

set.seed(123)
x <- sample(1:100, size = 500, replace = TRUE)

# into 5 groups, try to return similar group sizes
table(data_recode(x, split = "quantile", n_groups = 5))

# into 5 groups, try to return same range within groups
# i.e. 1-20, 21-40, 41-60, etc.
table(data_recode(x, split = "equal_range", range = 20))
}
