% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recode.R
\name{data_recode}
\alias{data_recode}
\alias{data_recode.numeric}
\alias{data_recode.data.frame}
\title{Recode old values of variables into new values}
\usage{
data_recode(x, ...)

\method{data_recode}{numeric}(x, recodes = NULL, verbose = TRUE, ...)

\method{data_recode}{data.frame}(
  x,
  recodes = NULL,
  force = FALSE,
  append = FALSE,
  select = NULL,
  exclude = NULL,
  ignore_case = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A data frame, numeric or character vector, or factor.}

\item{...}{not used.}

\item{recodes}{A list of named vectors, which indicate the recode pairs.
The \emph{names} of the list-elements (i.e. the left-hand side) represent the
old values, while the values of the list-elements indicate the values.
When recoding numeric vectors, element names have to be surrounded in
backticks. For example, \verb{recodes = list(\\}1\\verb{ = 0)} would recode all
\code{1} into \code{0} in a numeric vector.}

\item{verbose}{Toggle warnings.}

\item{select}{Variables that will be included when performing the required
tasks. Can be either
\itemize{
\item a variable specified as a literal variable name (e.g., \code{column_name}),
\item a string with the variable name (e.g., \code{"column_name"}), or a character
vector of variable names (e.g., \code{c("col1", "col2", "col3")}),
\item a formula with variable names (e.g., \code{~column_1 + column_2}),
\item a vector of positive integers, giving the positions counting from the left
(e.g. \code{1} or \code{c(1, 3, 5)}),
\item a vector of negative integers, giving the positions counting from the
right (e.g., \code{-1} or \code{-1:-3}),
\item or one of the following select-helpers: \code{starts_with("")}, \code{ends_with("")},
\code{contains("")}, a range using \code{:} or \code{regex("")}.
}

If \code{NULL}, selects all columns.}

\item{exclude}{See \code{select}, however, column names matched by the pattern
from \code{exclude} will be excluded instead of selected. If \code{NULL} (the default),
excludes no columns.}

\item{ignore_case}{Logical, if \code{TRUE} and when one of the select-helpers or
a regular expression is used in \code{select}, ignores lower/upper case in the
search pattern when matching against variable names.}
}
\value{
\code{x}, recoded into groups. By default \code{x} is numeric, unless \code{labels}
is specified. In this case, a factor is returned, where the factor levels
(i.e. recoded groups are labelled accordingly.
}
\description{
This functions recodes old values into new values and can be used to to
recode numeric or character vectors, or factors.
}
\examples{
set.seed(123)
x <- sample(c(1:4, NA), 20, TRUE)
table(x, useNA = "always")

out <- data_recode(x, list(`1` = 0, `2:3` = 1, `4` = 2))
out
table(out, useNA = "always")

out <- data_recode(x, list(`1` = 0, `2:3` = 1, `4` = 2, `NA` = 9))
out
table(out, useNA = "always")

out <- data_recode(x, list(`1` = 0, `2:3` = 1, `else` = 77))
out
table(out, useNA = "always")
}
