% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_findcols.R, R/data_relocate.R,
%   R/data_remove.R
\name{data_findcols}
\alias{data_findcols}
\alias{data_relocate}
\alias{data_reorder}
\alias{data_remove}
\title{Relocate (reorder) columns of a data frame}
\usage{
data_findcols(data, pattern = NULL, starts_with = NULL, ends_with = NULL, ...)

data_relocate(data, cols, before = NULL, after = NULL, safe = TRUE, ...)

data_reorder(data, cols, safe = TRUE, ...)

data_remove(data, pattern, ...)
}
\arguments{
\item{data}{A data frame to pivot.}

\item{pattern}{A regular expression (as character string), representing the
pattern to be matched in the in column names.}

\item{starts_with, ends_with}{Character string, containing the string to be
matched in the column names. \code{starts_with} finds matches at the beginning
of column names, \code{ends_with} finds matches at the end of column names.}

\item{...}{Other arguments passed to or from other functions.}

\item{cols}{A character vector indicating the names of the columns to move.}

\item{before, after}{Destination of columns. Supplying neither will move
columns to the left-hand side; specifying both is an error. Can be a
character vector, indicating the name of the destination column, or a
numeric value, indicating the index number of the destination column.
If \code{-1}, will be added before or after the last column.}

\item{safe}{If \code{TRUE}, will disregard non-existing columns.}
}
\value{
A data frame with reordered columns.
}
\description{
Relocate (reorder) columns of a data frame
}
\details{
\code{data_relocate()} will reorder columns to specific positions,
indicated by \code{before} or \code{after}. \code{data_reorder()} will always move
selected columns to the beginning of the data frame .
}
\examples{
# Find columns names by pattern
data_findcols(iris, starts_with = "Sepal")
data_findcols(iris, ends_with = "Width")
data_findcols(iris, pattern = "\\\\.")
data_findcols(iris, c("Petal.Width", "Sepal.Length"))

# Reorder columns
head(data_relocate(iris, cols = "Species", before = "Sepal.Length"))
head(data_relocate(iris, cols = "Species", before = "Sepal.Width"))
head(data_relocate(iris, cols = "Sepal.Width", after = "Species"))
# same as
head(data_relocate(iris, cols = "Sepal.Width", after = -1))

# reorder multiple columns
head(data_relocate(iris, cols = c("Species", "Petal.Length"), after = "Sepal.Width"))
# same as
head(data_relocate(iris, cols = c("Species", "Petal.Length"), after = 2))

# Reorder columns
head(data_reorder(iris, c("Species", "Sepal.Length")))
head(data_reorder(iris, c("Species", "dupa"))) # Safe for non-existing cols

# Remove columns
head(data_remove(iris, "Sepal.Length"))
}
\seealso{
\itemize{
\item Functions to rename stuff: \code{\link[=data_rename]{data_rename()}}, \code{\link[=data_rename_rows]{data_rename_rows()}}, \code{\link[=data_addprefix]{data_addprefix()}}, \code{\link[=data_addsuffix]{data_addsuffix()}}
\item Functions to reorder, find and remove columns: \code{\link[=data_findcols]{data_findcols()}}, \code{\link[=data_reorder]{data_reorder()}}, \code{\link[=data_relocate]{data_relocate()}}, \code{\link[=data_remove]{data_remove()}}
\item Functions to reshape, pivot or rotate dataframes: \code{\link[=data_to_long]{data_to_long()}}, \code{\link[=data_to_wide]{data_to_wide()}}, \code{\link[=data_rotate]{data_rotate()}}
\item Functions to rescale and reverse: \code{\link[=data_rescale]{data_rescale()}}, \code{\link[=data_reverse]{data_reverse()}}
\item Functions to standardize, normalize, rank-transform: \code{\link[=standardize]{standardize()}}, \code{\link[=normalize]{normalize()}}, \code{\link[=ranktransform]{ranktransform()}}, \code{\link[=winsorize]{winsorize()}}
\item Split, cut and merge dataframes: \code{\link[=data_partition]{data_partition()}}, \code{\link[=data_cut]{data_cut()}}, \code{\link[=data_match]{data_match()}}, \code{\link[=data_merge]{data_merge()}}
}
}
