% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\alias{data_summary.data.frame}
\title{Summarize data}
\usage{
data_summary(x, ...)

\method{data_summary}{data.frame}(x, ..., by = NULL, remove_na = FALSE, suffix = NULL)
}
\arguments{
\item{x}{A (grouped) data frame.}

\item{...}{One or more named expressions that define the new variable name
and the function to compute the summary statistic. Example:
\code{mean_sepal_width = mean(Sepal.Width)}. The expression can also be provided
as a character string, e.g. \code{"mean_sepal_width = mean(Sepal.Width)"}. The
summary function \code{n()} can be used to count the number of observations.}

\item{by}{Optional character string, indicating the names of one or more
variables in the data frame. If supplied, the data will be split by these
variables and summary statistics will be computed for each group.}

\item{remove_na}{Logical. If \code{TRUE}, missing values are omitted from the
grouping variable. If \code{FALSE} (default), missing values are included as a
level in the grouping variable.}

\item{suffix}{Optional character vector or a list of character vectors,
indicating the suffixes to be added to the new variable names. This is useful
when the summary function returns more than one value (e.g., \code{quantile()}).
If a list, it should have the same length as the number of expressions in
\code{...}. The new column names are a combination of the left-hand side (i.e.,
the name) of the expression and the related suffixes. See 'Examples'.}
}
\value{
A data frame with the requested summary statistics.
}
\description{
This function can be used to compute summary statistics for a
data frame or a matrix.
}
\examples{
data(iris)
data_summary(iris, MW = mean(Sepal.Width), SD = sd(Sepal.Width))
data_summary(
  iris,
  MW = mean(Sepal.Width),
  SD = sd(Sepal.Width),
  by = "Species"
)

# same as
d <- data_group(iris, "Species")
data_summary(d, MW = mean(Sepal.Width), SD = sd(Sepal.Width))

# multiple groups
data(mtcars)
data_summary(mtcars, MW = mean(mpg), SD = sd(mpg), by = c("am", "gear"))

# expressions can also be supplied as character strings
data_summary(mtcars, "MW = mean(mpg)", "SD = sd(mpg)", by = c("am", "gear"))

# count observations within groups
data_summary(mtcars, observations = n(), by = c("am", "gear"))

# first and last observations of "mpg" within groups
data_summary(
  mtcars,
  first = mpg[1],
  last = mpg[length(mpg)],
  by = c("am", "gear")
)

# allow more than one-column-summaries for expressions
d <- data.frame(
  x = rnorm(100, 1, 1),
  y = rnorm(100, 2, 2),
  groups = rep(1:4, each = 25)
)

# same suffix for each new column created by the expression
data_summary(
  d,
  quant_x = quantile(x, c(0.25, 0.75)),
  quant_y = quantile(y, c(0.25, 0.75)),
  suffix = c("Q1", "Q3")
)

# different summary lengths, requires a list of suffixes
data_summary(
  d,
  quant_x = quantile(x, c(0.25, 0.75)),
  quant_y = quantile(y, c(0.25, 0.5, 0.75)),
  suffix = list(c("Q1", "Q3"), c("_Q1", "_Q2", "_Q3"))
)

# also works with grouped data, and multiple suffixes
data_summary(
  d,
  quant_x = quantile(x, c(0.25, 0.75)),
  quant_y = quantile(y, c(0.1, 0.9)),
  by = "groups",
  suffix = list(c("Q1", "Q3"), c("10perc", "90perc"))
)

data_summary(
  d,
  quant_x = quantile(x, c(0.25, 0.75)),
  quant_y = quantile(y, c(0.25, 0.5, 0.75)),
  suffix = list(c("Q1", "Q3"), c("_Q1", "_Q2", "_Q3")),
  by = "groups"
)

}
