% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_unique.R
\name{data_unique}
\alias{data_unique}
\title{Keep only one row from all with duplicated IDs}
\usage{
data_unique(data, id, method = "best")
}
\arguments{
\item{data}{The data frame.}

\item{id}{The ID variable for which to check for duplicates.}

\item{method}{The method to be used for duplicate selection,
either "best" (the default), "first", or "last".}
}
\value{
A dataframe, containing only the chosen duplicates.
}
\description{
From all rows with at least one duplicated ID,
keep only one. Methods for selecting the duplicated row are
either the first duplicate, the last duplicate, or the "best"
duplicate (default), based on the duplicate with the fewer
number of NA values. In case of ties, it picks the first
duplicate, as it is the one most likely to be valid and
authentic, given practice effects.
}
\examples{
df1 <- data.frame(
   id = c(1, 2, 3, 1, 3),
   item1 = c(NA, 1, 1, 2, 3),
   item2 = c(NA, 1, 1, 2, 3),
   item3 = c(NA, 1, 1, 2, 3)
)

data_unique(df1, id = "id")
}
