% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{distribution_mode}
\alias{distribution_mode}
\title{Compute mode for a statistical distribution}
\usage{
distribution_mode(x)
}
\arguments{
\item{x}{An atomic vector, a list, or a data frame.}
}
\value{
The value that appears most frequently in the provided data.
The returned data structure will be the same as the entered one.
}
\description{
Compute mode for a statistical distribution
}
\examples{

distribution_mode(c(1, 2, 3, 3, 4, 5))
distribution_mode(c(1.5, 2.3, 3.7, 3.7, 4.0, 5))

}
\seealso{
For continuous variables, the
\strong{Highest Maximum a Posteriori probability estimate (MAP)} may be
a more useful way to estimate the most commonly-observed value
than the mode. See \code{\link[bayestestR:map_estimate]{bayestestR::map_estimate()}}.
}
